requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10

